Split huge class into pieces.
