+++ src/test/java/org/carrot2/elasticsearch/MultithreadedClusteringIT.java
@@ -7,8 +7,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import org.assertj.core.api.Assertions;
-import org.carrot2.elasticsearch.ClusteringAction.ClusteringActionRequestBuilder;
-import org.carrot2.elasticsearch.ClusteringAction.ClusteringActionResponse;
 import org.elasticsearch.client.Client;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.fetch.subphase.highlight.HighlightBuilder;
