+++ src/test/java/org/carrot2/elasticsearch/SampleIndexTestCase.java
@@ -16,8 +16,7 @@ import org.apache.http.HttpHeaders;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
 import org.assertj.core.api.Assertions;
-import org.carrot2.elasticsearch.ClusteringAction.ClusteringActionResponse;
-import org.carrot2.elasticsearch.ClusteringAction.ClusteringActionResponse.Fields;
+import org.carrot2.elasticsearch.ClusteringActionResponse.Fields;
 import org.carrot2.language.LanguageComponentsLoader;
 import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
 import org.elasticsearch.action.bulk.BulkRequestBuilder;
